Description

Unveil

Unveil.combinecv — Method

WORK IN PROGRESS Combine multiple CV maps into one cube. Use this script in a julia terminal with : julia>Unveil.combinecv()

Unveil.compmethod_stcfct — Method

Comparing PCA and SWO method by printing exponant of the structure functions with the order on the same figure

Unveil.convpca — Method

convpca(VARFILEPATH)

Produce calculations to find the PCA convergence criteria based on the matrix projection from PCA. No PCA reconstructed cube will be saved. A '.txt' file should be used accordingly as an input (see models inside folders '/varfiles/convpca.txt').

OUTPUTS : A plots with every moments of the projection matrix + the metric (see the doc). Also add a .dat file with moments, metric and number of PC used for each.

Use this script in a julia terminal with : julia>Unveil.convpca(VARFILEPATH)

Unveil.cv — Method

cv(VARFILEPATH)

Calculate the CV from a cube given in input. A '.txt' file should be used accordingly as an input (see models inside folders '/varfiles/cv.txt').

Use this script in a julia terminal with : julia>Unveil.cv(VARFILEPATH)

Unveil.cvcvi — Method

cvcvi(VARFILEPATH)

Calculate the CV and CVI from a cube given in input. A '.txt' file should be used accordingly as an input (see models inside folders '/varfiles/cvcvi.txt').

Use this script in a julia terminal with : julia>Unveil.cvcvi(VARFILEPATH)

Unveil.cvi — Method

cvi(VARFILEPATH)

Calculate the CVI from a CV map. A '.txt' file should be used accordingly as an input (see models inside folders '/varfiles/cvi.txt').

INPUT : path to the variable '.txt' file OUTPUTS : Save the CV fits,

Use this script in a julia terminal with : julia>Unveil.cvi(VARFILEPATH)

Unveil.multipca — Method

multipca(VARFILEPATH)

Use multiple PCA processes on a cube and with multiple numbers of PCs given as input. A '.txt' file should be used accordingly as an input (see models inside folders '/varfiles/multipca.txt').

OUTPUTS : One file per cube reconstructed with one of the number of PC given as an input.

Use this script in a julia terminal with : julia>Unveil.multipca(VARFILEPATH)

Unveil.pca — Method

pca(VARFILEPATH)

Use a PCA (Principal Component Analysis) process on a cube and with N PCs given as input. A '.txt' file should be used accordingly as an input (see models inside folders '/varfiles/pca.txt').

INPUTS : path to the variable '.txt' file OUTPUTS : save a cube reconstructed by the number of PC asked in the '.txt' file

Use this function in a julia terminal with : julia> Unveil.pca(VARFILEPATH)

Unveil.structure_functions — Method

structure_functions(VARFILEPATH)

Compute the structure functions $S_p(l)$ of a cvi cube (following definition in Kritsuk+2007). A '.txt' file should be used accordingly as an input (see models inside folders '/varfiles/structure_functions.txt'). The cube given as input should be a CVI (use function Unveil.cvi if needed). Prefer a cube with rotations of every lags than azimutal average, like that : (Pixel positions,angles,lag).

OUTPUTS : One figure with $S_p(l)$ vs $S_3(l)$, same with fit, and exponants of $S_p(l)$ function of p. Also a .dat file with values of the exponants. Use this function in a julia terminal with : julia> Unveil.structure_functions(VARFILEPATH)

Unveil.swo — Method

swo(VARFILEPATH)

Use a SWO (Spectral Window Optimisation) process on a cube. A '.txt' file should be used accordingly as an input (see models inside folders '/varfiles/swo.txt').

Use this function in a julia terminal with : julia> Unveil.swo(VARFILEPATH)

Data_preparation

Dataprep.addblank — Method

addblank(data,missing_arr,blank,data_dimension)

Reconstruct a data with missing values after they where deleted. Usually deleted in order to compute a PCA. Can reconstruct 3D cubes (e.g. PPV) as well as 2D maps (PP). In the first case, the input missingarr should be in 2D, while in the second case it should be in 1D. The input missingarr is produces with the function Datapreparation.pcaprep. The input data should be a 1D vector. Data_dimension has to be equal to the dimension you want to obtain. To reproduce a 3D cube, it should have blank values at the exact same positions across the 3rd dimension.

Dataprep.blank_equal — Method

blank_equal(array,oldvalue,newvalue)

Change the values equals to a value by another one in an array. This is useful if change of blanking is desired. You have to use the function "replace_nantoblank" if the old value is NaN.

Dataprep.blank_inf — Method

blank_inf(array,oldvalue,newvalue)

Change the values tinier than an old value to a new value in an array. This is useful if a change of blanking is desired

Dataprep.boolmatrix_missing — Method

boolmatrix_missing(array)

Return a 1D and a 2D boolean matrices indicating the indexes of missing values in array. Array should be a 2 dimension data (pixels*spectra for example)

Dataprep.delete_allnotvalue — Method

delete_allnotvalue(data,blank)

Delete from data all value which are not a valid value, e.g. missing, NaN and blank.

Dataprep.deletemissing — Method

deletemissing(data,missing1D)

Delete all rows where there is at least one missing value using a boolean vector of the missing values.

Dataprep.directory_prep — Method

directory_prep(PATHTOSAVE::String)

Construct a new directory to save plots of the convergence criteria (Convergence_Criteria) process and another to save the fits (Data).

Dataprep.pca_prep — Method

pca_prep(arr,arraydimension)

Prepare data (given in 2D (pv) or 3D (ppv)) in order to conduct PCA on them by deleting missing values and reshape in 2D.

Will also check if any missing value still exist in the data, showing that the data are not regularly blanked. Return data in 2 dimensions without missing value, a 1D boolean matrix to check if any missing value still exist in the data (if 0 element then no missing value) and a 2D matrix with booleans corresponding to missing values in the dataset (used to reconstruct the data with the missing values).

Dataprep.permcolrow — Method

permcolrow(arr)

Return the array with the first and second dimensions permuted. 2D and 3D arrays accepted.

Dataprep.read_dat — Method

read_dat(DATFILEPATH ; com='#')

Return values of the .dat file given as input. By default, will consider as comment rows starting with a #. Can be changed.

Dataprep.read_dim — Method

read_dim(arr)

Return a vector with dimensions of the data.

Dataprep.read_fits_cvi — Method

read_fits_cvi(path ; check = true)

Return the cube, its dimension and its header. Use to read CVI cube, with this order : Pixel position, Rotation, LAG.

Dataprep.read_fits_pp — Method

read_fits_pp(path)

Read a 2D fits (pixel*pixel). Return values (matrix), header and dimensions.

Dataprep.read_fits_ppv — Method

read_fits_ppv(path,vel_units ; check = true)

Read data of a PPV fits from its path, test if the fits is conform, then return : [1] Data in an array [2] Range of velocity in an array (in km/s) [3] Dimensions of the data (one vector with each element a dimension) [4] The velocity resolution (in km/s) [5] The header of the fits If the header does not contain any third dimension indication, the function will still return the data, the dimension and the header. Velocities need to be in the third dimension. This function conduct a sanity check by default. Give check = false if you don't want the test being conducted.

Dataprep.read_var_files — Method

read_var_files(varfile_path)

A function to read and import values of a variable files. The variable files can be anywhere in the machine, but preferentially localised in the "var_file" folders.

Dataprep.regular_blanking — Method

regular_blanking(data)

Check if the data of a fits file contains any irregular blanking e.g. a spectra with one blanking value (or more, but less than the spectra size).

Dataprep.replace_blanktomissing — Method

replace_blanktomissing(arr,blanktomissing)

Replace specific values (blanktomissing, generally blank value) by missing value in an array.

Dataprep.replace_inf_in_nan — Method

replace_inf_in_nan(arr)

Replace the Inf values by NaN values. Useful when searching for max and min values in the array.

Dataprep.replace_missingtoblank — Method

replace_missingtoblank(arr,missingtoblank)

Replace missing values by a 'missingtoblank' value in an array (generally a blank value).

Dataprep.replace_nantoblank — Method

replace_nantoblank(data,blank)

Replace the NaN values in an array by a new blank value.

Dataprep.replace_nantomissing — Method

replace_nantomissing(data)

Replace the NaN values in an array by missing value.

Dataprep.replace_nosignal — Method

function replace_nosignal(cube,DATADIMENSION,VELOCITYVECTOR,BLANK,SIGMAMAP)

Replace spectra without signal by blank value. We recover the spectra without signal by compute the intensity integration of each of them. If it is less than 2 times the dispersion of the noise, then it is blanked. SIGMAMAP is the 2D map of each rms noise of your cube ; use Dataanalysis.rmscube(cube,NOISECAN)[1] to compute it.

Dataprep.write_dat — Method

write_dat(matrix,PATHTOSAVE,NAME ; more=[""], overwrite=false)

Write matrix in a .dat file. Each column of the matrix is writen as new column in the file. Delimiters are spaces by default. The option overwrite is false by default. If want to overwrite file with the same name as given in input, change it to true. If want to add more entries on the header of the dat, use the option 'more'. Will be added as commentaries with a # at the begining of the line.

Dataprep.write_fits — Method

write_fits(fitstocopy::String,newname::String,pathtosave::String,datatosave,datadim,blank ; finished=true)

Create a fits with its own header based on 'fitstocopy', and with the name 'newfits'. DO NOT ADD THE EXTENSION (e.g ".fits") FOR THE NAME OF THE NEWFITS). Need to give the blanking value. The fits can be constructed by data with two dimensions or three dimensions. Variable 'finished' is by default true, so the function will by default close the new fits file at the end of the execution of this function. If 'finished' is false, the function will not close the fits file, allowing it to be modified again in another function (createcvifits for example). If want to add more entries on the header, use the option 'more'. Alternate between the name printed in the header and the value. Example, if want to add an entry for 'imsize' equals to 50, and an entry for 'powerlaw' equals to -3, write : more=["imsize","50","powerlaw","-3"]. The option overwrite is false by default. If want to overwrite file with the same name as given in input, change it to true.

PCA

PCA.pca — Method

    pca(data2D,pc::Integer,path::String ; percent=1.0)

Produce a Principal Component Analysis on a data, using 'pc' number of principal component. 'percent' gives the wanted percentage of variance reconstructed (default=1). 'path' is the path where results will be saved (in 'bin' temporary files, named as mmapYt.bin, mmapMproj.bin and mmapMmean.bin) Data are converted in an Array{Float64,2} to do the PCA ( type as Union{Missing} will produce an error). Data should be given in 2D (pixels per velocities). Will return the path to the projections of the data2D on the PC axis, the mean of these projections, the path to the matrix composed of the principal components in columns and the percentage of explained variance. These results will allow to reconstruct the cube with a given number of PC (using pcanomorecalc).

PCA.pca — Method

    pca(data2D,pc::Integer ; percent=1.0)

Produce a Principal Component Analysis on a data, using 'pc' number of principal component. 'percent' gives the wanted percentage of variance reconstructed (default=1). Data are converted in an Array{Float64,2} to do the PCA ( type as Union{Missing} will produce an error). Data should be given in 2D (pixels per velocities). Will return the pca fit M (a type created by the MultivariateStats package), the matrix composed of the principal components in columns, the percentage of explained variance and the data reconstructed with 'pc' number of principal component.

PCA.pca_nomorecalc — Method

    pca_nomorecalc(M,Yt,pcfirst,pclast,Datadim)

Reconstruct data with a certain number of PC, without redo all of the PCA calculation. Necessary to do a first pca calculation to run this one, because needed M (a new type related to PCA constructed by MultivariateStats package) and Yt (matrix with each column a PC). Datadim should be a tuple with two values (PV cube, 2D) or three (PPV cube, 3D).

PCA.pca_nomorecalc — Method

    pca_nomorecalc(Mproj_path::String,Yt_path::String,Mmean_path::String,pcfirst,pclast,datadim)

Same as pca_nomorecalc(M,Yt,pcfirst,pclast,Datadim) but using here the path to the temporary files containing the informations (M and Yt). To be used after pca(data2D,pc::Integer,path::String ; percent=1.0) in order to produce a reconstructed cube.

SWO

SWO.bestsnr — Method

bestsnr(cubesource,DATADIMENSION_NOMISSING,VELOCITYVECTOR,NOISECAN)

Method SWO (Spectral Window Optimisation). For each spectra of the cubesource, look for the maximum SNR given by :

\[SNR(v_\text{i})=\frac{\sum_{\text{i}=1}^\text{m}T_\text{i}}{\sqrt{m\sigma_{1:m}}}\]

In this equation, $T_\text{i}$ is the intensity at velocity channel $v_\text{i}$, and $\sigma_{1:m}$ the dispersion computed between velocity channels 1 to m. The SNR is computed in increasing velocity channels, then in decreasing velocity channels. Each maxima of these two computations will give one window limit containing the emission.

INPUT : cubesource, the cube whom you want to compute the SWO. It should be in 2D (PxP,V), as DATADIMENSION_NOMISSING, which gives the total dimensions of the cube. VELOCITYVECTOR is the velocity vector, computed and given as output of the function Dataprep.read_fits_ppv. NOISECAN are some of the velocity channels of the cube where there is only noise. Used to compute the dispersion of the noise of cubesource.

OUTPUT : [1] Cubesource with SWO mask applied (inside mask windows are equal to cubesource, outside is 0) [2] Mask of SWO method computed on cubesource

CVI

CVI.construct_cvimap! — Method

construct_cvimap!(xyarr,Lag::Int64,mapdim,cvi_averaged::Array{Union{Missing, Float64}},cvi_allangle::Array{Union{Missing,Float64},3}; diff="relative",keepmissing=true)

Will construct a cvi map using preallocated array. Need an array preallocated for the storage of the average on all cvi angle (cviaveraged), and another for the cvi calculated with all angle and lag (cviallangle). The cvi map is constructed by taking the mean of all rotations of the cv increment calculation at each pixel. The Lag is the increment. xyarr have to be in 2D (pixel*pixel). Mapdim is the dimension of your 2Dmap. The differences can be absolute or relative.

CVI.construct_cvimap! — Method

construct_cvimap!(xyarr,Lag::Vector{Int64},mapdim,cvi_averaged_alllag::Array{Union{Missing, Float64},2},cvi_allangle_alllag::Array{Union{Missing, Float64}, 3},cvi_allangle::Array{Union{Missing,Float64},3}; diff="relative",keepmissing=true)

Will construct a cvi map using preallocated array. Need an array preallocated for the storage of the average on all cvi angle (cvimean), and another for the cvi calculated with all angle and lag (cviallanglealllag). The cvi map is constructed by taking the mean of all rotations of the cv increment calculation at each pixel. The Lag is the increment. xyarr have to be in 2D (pixel*pixel). Mapdim is the dimension of your 2Dmap. The differences can be absolute or relative.

CVI.construct_cvimap — Method

construct_cvimap(cvmap,Lag::Int64,mapdim; diff="relative",keepmissing=true)

Same as construct_cvimap if Lag is a Int64 of one lag. Mapdim is the dimension of your 2Dmap. The differences can be absolute or relative.

CVI.construct_cvimap — Method

construct_cvimap(cvmap,Lag::Vector{Int64},mapdim; diff="relative",keepmissing=true,BLANK=-1000)

Construct a Centroid Velocity Increment map based on a 'cvmap'. The cvi map is constructed by taking the mean of all rotations of the Centroid Velocity Increment calculation at each pixel. The Lag is the increment. xyarr have to be in 2D (pixel*pixel). Mapdim is the dimension of your 2Dmap. The differences can be absolute or relative.

CVI.cv_increment! — Method

cv_increment!(xyarr,Lag::Int64,blank,cvimap_lag::Array{Union{Missing,Float64},2})

Same as (xyarr,Lag::Int64,nangle; diff="relative",periodic=false) but on preallocated array (array produced outside this function). To produce the array, use : cvi_allangle = Array{Union{Missing,Float64},3}(undef,DataDimension[1],DataDimension[2],nangle).

CVI.cv_increment! — Method

cv_increment!(xyarr,Lag::Vector{Int64},nangle,cvi_allangle_alllag::Array{Union{Missing, Float64}, 3},cvi_allangle ; diff="relative",periodic=false)

Same as cvincrement(xyarr,Lag::Vector{Int64},nangle; diff="relative",periodic=false) but on preallocated array (array produced outside this function). To produce the array, use : cviallangle = Array{Union{Missing,Float64},3}(undef,DataDimension[1],DataDimension[2],nangle) and cviallanglealllag = Array{Union{Missing,Float64},3}(undef,DataDimension[1]*DataDimension[2],nangle,size(Lag)[1])

CVI.cv_increment — Method

cv_increment(xyarr,Lag::Int64,nangle; diff="relative",periodic=false)

Compute the centroid velocity increment of xyarr at one Lag. Nangle is the number of angle using to compute the differences (it's a value in the parameter file, equal to 192). Diff (default relative) is for differences between two pixels : absolute or relative. Periodic=true (default=false) is for working on periodic data (from simulations like fbm). The returned array will have the first two dimensions equal to the size of the map, and the third dimension is the cvi computed at each angle.

CVI.cv_increment — Method

cv_increment(xyarr,Lag::Vector{Int64},nangle; diff="relative",periodic=false)

Compute the centroid velocity increment of xyarr at multiple Lag values. Nangle is the number of angle using to compute the differences (it's a value in the parameter file, equal to 192). Diff (default relative) is for differences between two pixels : absolute or relative. Periodic=true (default=false) is for working on periodic data (from simulations like fbm). The returned array will have the first dimension equal to the size of the map (pixel square), the second dimension is the cvi computed at each angle, and the third dimension is for each value of Lag.

CVI.moment_one — Method

moment_one(yarr,xarr)

Compute the first moment order of the array xarr weighted by yarr. Centroid = int(T*v)dv/int(T)dv

CVI.moment_one_field — Method

moment_one_field(arr,velvector)

Return the first moment order of all pixels in an entire field, weighted by velvector. Arr can be a 3D array (ppv) or a 2D array (pv). The function replace all missing values by

Graphical representations

Graphic.StcFct — Method

StcFct()

Plot structure functions of orders p in function of the structure function of order 3. If add is true, the plot will be added on the actual plot. The figure can be saved with the number of PC used during the process in the name by changing pcfinal.

Graphic.StcFctExponent — Method

StcFctExponent(zeta,third_order_zeta,OrderP,xlim,ylim,labs,data_name_title,directories_name; add=false,save=true,pcfinal=0)

Plot the exponent of the structure function power-law fit in function of the order p of the structure function. If add is true, the plot will be added on the actual plot. Also add a y=1/3x line on the plot. The figure can be saved with the number of PC used during the process in the name by changing pcfinal.

Graphic.StcFctWithFit — Method

StcFctWithFit()

Plot structure functions of orders p in function of the structure function of order 3 with the fit used to obtain the power-law exponent. If add is true, the plot will be added on the actual plot. The figure can be saved with the number of PC used during the process in the name by changing pcfinal.

Functions for Data Analysis

Analysis.fourmoments — Method

fourmoments(cube;dim=2)

Compute the first four moments order of a given cube of dimension 'dim'. If 2D PV cube, will compute the moments on each row. Return the four moments. They are Vector of dimension=size(cube)[1]

Analysis.metricOW — Method

metricOW(mom1,mom2,mom3,mom4,dv,SIGMAT)

Calculate the following metric on each values of mom1,2,3 and 4: sqrt((mom1)2+((mom2-SIGMAT))2+(mom3)2+(mom4-3)2)

Analysis.metricPCA — Method

metricPCA(mom1,mom2,mom3,mom4,dv)

Calculate the following metric on each values of mom1,2,3 and 4: sqrt((mom1/dv)2+(mom2/dv)2+(mom3)2+(mom4-3)2)

Analysis.rms_cube — Method

rms_cube(cube,can)

Compute the rms on velocity canal given as input of a cube. Return a 2D map with given rms on each pixel and the averaged rms accross the map.

WORK IN PROGRESS : Function for Structure functions computations

Structure_functions.fit_fctsct — Method

fit_fctstruct(xdata,ydata,y0,confidinterv)

Fit the model ydata=A*(xdata)^B. Return in first B, then A.

Structure_functions.xhi_fct_p — Method

xhi_fct_p(pvec,struct_lag,y0)

Fit the model y=A*(x)^B for multiple order and lag of structure functions.