Description

Unveil

Unveil.combinecv — Method

WORK IN PROGRESS Combine multiple CV maps into one cube. Use this script in a julia terminal with : julia>Unveil.combinecv()

Unveil.convpca — Method

convpca(VARFILEPATH)

Produce calculations to find the PCA convergence criteria based on the matrix projection from PCA. No PCA reconstructed cube will be saved. A '.txt' file should be used accordingly as an input (see models inside folders '/varfiles/convpca.txt').

OUTPUTS : A plots with every moments of the projection matrix + the metric (see the doc). Also add a .dat file with moments, metric and number of PC used for each.

Use this script in a julia terminal with : julia>Unveil.convpca(VARFILEPATH)

Unveil.convswo — Method

convswo(VARFILEPATH)

Look for the best interval to integrate the spectra for the window optimisation process. A '.txt' file should be used accordingly as an input (see models inside folders '/varfiles/convswo.txt').

Use this script in a julia terminal with : julia>Unveil.convswo(VARFILEPATH)

Unveil.cv — Method

cv(VARFILEPATH)

Calculate the CV from a cube given in input. A '.txt' file should be used accordingly as an input (see models inside folders '/varfiles/cv.txt').

Use this script in a julia terminal with : julia>Unveil.cv(VARFILEPATH)

Unveil.cvcvi — Method

cvcvi(VARFILEPATH)

Calculate the CV and CVI from a cube given in input. A '.txt' file should be used accordingly as an input (see models inside folders '/varfiles/cvcvi.txt').

Use this script in a julia terminal with : julia>Unveil.cvcvi(VARFILEPATH)

Unveil.cvi — Method

cvi(VARFILEPATH)

Calculate the CVI from a CV map. A '.txt' file should be used accordingly as an input (see models inside folders '/varfiles/cvi.txt').

INPUT : path to the variable '.txt' file OUTPUTS : Save the CV fits,

Use this script in a julia terminal with : julia>Unveil.cvi(VARFILEPATH)

Unveil.multipca — Method

multipca(VARFILEPATH)

Use multiple PCA processes on a cube and with multiple numbers of PCs given as input. A '.txt' file should be used accordingly as an input (see models inside folders '/varfiles/multipca.txt').

OUTPUTS : One file per cube reconstructed with one of the number of PC given as an input.

Use this script in a julia terminal with : julia>Unveil.multipca(VARFILEPATH)

Unveil.pca — Method

pca(VARFILEPATH)

Use a PCA (Principal Component Analysis) process on a cube and with N PCs given as input. A '.txt' file should be used accordingly as an input (see models inside folders '/varfiles/pca.txt').

INPUTS : path to the variable '.txt' file OUTPUTS : save a cube reconstructed by the number of PC asked in the '.txt' file

Use this function in a julia terminal with : julia> Unveil.pca(VARFILEPATH)

Unveil.swo — Method

swo(VARFILEPATH)

Use a SWO (Spectral Window Optimisation) process on a cube. A '.txt' file should be used accordingly as an input (see models inside folders '/varfiles/swo.txt').

Use this function in a julia terminal with : julia> Unveil.swo(VARFILEPATH)

Data_preparation

Data_preparation.addblank — Method

addblank(data,missing_arr,blank,data_dimension)

Reconstruct a data with missing values after they where deleted. Usually deleted in order to compute a PCA. Can reconstruct 3D cubes (e.g. PPV) as well as 2D maps (PP). In the first case, the input missingarr should be in 2D, while in the second case it should be in 1D. The input missingarr is produces with the function Datapreparation.pcaprep. The input data should be a 1D vector. Data_dimension has to be equal to the dimension you want to obtain. To reproduce a 3D cube, it should have blank values at the exact same positions across the 3rd dimension.

Data_preparation.addmask — Method

addmask(cube,colorscale,nbmask,data_dimension,delta_xvec,delta_yvec)

Blank all the pixels in the third dimension selected manually by the user (using a contour). Can do multiple mask.

Data_preparation.allmissing — Method

allmissing(data)

Return true if the data contain any missing value.

Data_preparation.blank_equal — Method

blank_equal(array,oldvalue,newvalue)

Change the values equals to a value by another one in an array. This is useful if change of blanking is desired. You have to use the function "replace_nantoblank" if the old value is NaN.

Data_preparation.blank_inf — Method

blank_inf(array,oldvalue,newvalue)

Change the values tinier than an old value to a new value in an array. This is useful if a change of blanking is desired

Data_preparation.blank_inf_abs — Method

blank_inf_abs(array,oldvalue,newvalue)

Change the values tinier than an old value (its absolute value) to a new value in an array. This is useful if a change of blanking is desired

Data_preparation.blank_snr — Method

blank_snr(data2D,NoiseCan,thresh)

Blank all the spectra which have a signal to noise ratio inferior than a threshold. Work on 2D data, the first dimension giving the pixel position and the second dimension the spectra referring to this position.

Data_preparation.blank_sup — Method

blank_sup(array,oldvalue,newvalue)

Change the values bigger than an old value to a new value of an array . This is useful if a change of blanking is desired

Data_preparation.boolmatrix_missing — Method

boolmatrix_missing(array)

Return a 1D and a 2D boolean matrices indicating the indexes of missing values in array. Array should be a 2 dimension data (pixels*spectra for example)

Data_preparation.clear_var — Method

clear_var(var1,var2,var3,var4)

Replace all variable in input by 0 to "free" space.

Data_preparation.coordtopix — Method

coordtopix(header,coord,unit,ax ; off=true)

Do the conversion between physical coordinate and pixel number thanks to the header of the data. Works only for small-size fields. Need the unit of the coordinate given (arcsec,arcmin,rad,deg only) and the axis. Consider by default that the coordinate given is an offset (place at the center of the image). The values of CRVAL1... are considered in deg.

Data_preparation.delete_allnotvalue — Method

delete_allnotvalue(data,blank)

Delete from data all value which are not a valid value, e.g. missing, NaN and blank.

Data_preparation.deletemissing — Method

deletemissing(data,missing1D)

Delete all rows where there is at least one missing value using a boolean vector of the missing values.

Data_preparation.directory_prep — Method

directory_prep(PATHTOSAVE::String)

Construct a new directory to save plots of the convergence criteria (Convergence_Criteria) process and another to save the fits (Data).

Data_preparation.lookingfor_interval — Method

lookingfor_interval(arr,value,sigma)

Return the array (1D) cut at [value-sigma:value+sigma]. Return also the indices of these values.

Data_preparation.noise_channels — Method

noise_channels()

Construct a vector with two subranges.

Data_preparation.pca_prep — Method

pca_prep(arr,arraydimension)

Prepare data (given in 2D (pv) or 3D (ppv)) in order to conduct PCA on them by deleting missing values and reshape in 2D.

Will also check if any missing value still exist in the data, showing that the data are not regularly blanked. Return data in 2 dimensions without missing value, a 1D boolean matrix to check if any missing value still exist in the data (if 0 element then no missing value) and a 2D matrix with booleans corresponding to missing values in the dataset (used to reconstruct the data with the missing values).

Data_preparation.permcolrow — Method

permcolrow(arr)

Return the array with the first and second dimensions permuted. 2D and 3D arrays accepted.

Data_preparation.pixtocoord — Method

pixtocoord(header,unit)

Do the conversion between pixel number and physical coordinate thanks to the header of the data. Return the X and Y coordinates in the physical coordinates you want, then the X and Y coordinate centered at the middle of the plot.

Data_preparation.read_dim — Method

read_dim(arr)

Return a vector with dimensions of the data.

Data_preparation.read_fits_pp — Method

read_fits_pp(path)

Read a 2D fits (pixel*pixel). Return values (matrix), header and dimensions.

Data_preparation.read_fits_ppf — Method

read_fits_ppf(field ; path="none", check = true) WIP

Read data of a PPF fits from its path, test if the fits is conform, then return : [1] Data in an array [2] Range of velocity in an array (in km/s) [3] X coordinate in arcsecond [4] Y coordinate in arcsecond [5] X coordinates centered in the middle of the image [6] Y coordinates centered in the middle of the image [7] Dimensions of the data (one vector with each element a dimension) [8] The velocity resolution (in km/s) [9] The header of the fits [10] The path to the fits Make the conversion of frequency into velocity. If the header does not contain any third dimension indication, the function will still return the data, the dimension and the header. Velocities need to be in the third dimension. This function conduct a sanity check by default. Give check = false if you don't want the test being conducted.

Data_preparation.read_fits_ppv — Method

read_fits_ppv(path,vel_units ; check = true)

Read data of a PPV fits from its path, test if the fits is conform, then return : [1] Data in an array [2] Range of velocity in an array (in km/s) [3] Dimensions of the data (one vector with each element a dimension) [4] The velocity resolution (in km/s) [5] The header of the fits If the header does not contain any third dimension indication, the function will still return the data, the dimension and the header. Velocities need to be in the third dimension. This function conduct a sanity check by default. Give check = false if you don't want the test being conducted.

Data_preparation.read_fits_spect — Method

read_fits_spect(path,vel_units )

Read data of a spectra PV fits from the path of the field and return : [1] an array of the data [2] an array of the range of velocity (in km/s) [3] the velocity resolution (in km/s) [4] the header of the fits [5] the file of the fits

Velocities need to be in the second dimension. By default, the path of the fits is "none", because we usually use local path listed in the function data_path.

Data_preparation.read_fits_velocityvector — Method

read_fits_velocityvector(path,vel_units)

Reconstruct the velocity vector of a fits file. Have to specify the units of the velocity vector of the fits.

Data_preparation.read_var_files — Method

read_var_files(varfile_path)

A function to read and import values of a variable files. The variable files can be anywhere in the machine, but preferentially localised in the "var_file" folders.

Data_preparation.regular_blanking — Method

regular_blanking(data)

Check if the data of a fits file contains any irregular blanking e.g. a spectra with one blanking value (or more, but less than the spectra size).

Data_preparation.replace_blanktomissing — Method

replace_blanktomissing(arr,blanktomissing)

Replace specific values (blanktomissing, generally blank value) by missing value in an array.

Data_preparation.replace_inf_in_nan — Method

replace_inf_in_nan(arr)

Replace the Inf values by NaN values. Useful when searching for max and min values in the array.

Data_preparation.replace_missingtoblank — Method

replace_missingtoblank(arr,missingtoblank)

Replace missing values by a 'missingtoblank' value in an array (generally a blank value).

Data_preparation.replace_nantoblank — Method

replace_nantoblank(data,blank)

Replace the NaN values in an array by a new blank value.

Data_preparation.replace_nantomissing — Method

replace_nantomissing(data)

Replace the NaN values in an array by missing value.

Data_preparation.valid_header — Method

    valid_header(header)

Check if the header of a fits file contains a blank, a bzero, and a bscale values.

If not, return an error.

Data_preparation.wait_for_key — Method

wait_for_key(; prompt = "Press any key", io = stdin))

Pause the script until the user press a key. The text prompt can be changed.

Data_preparation.write_cv_fits — Method

write_cv_fits(fitstocopy::String,newname::String,data,datadim,blank ; pc=0,varpercent=100)

Create a fits file used for cv reconstructed data. Contains on its header the number of pc used (default=0), the percentage of variance explained (default=100), the lag and the angle. Need a fits to copy its header entries.

Data_preparation.write_cvi_fits — Method

write_cvi_fits(lag,fitstocopy::String,newname::String,data,datadim,blank; pc=0,varpercent=100)

Create a fits used for cvi map. Contains on its header the lag used to do the cvi. Possibility to add in the header the number of PC used if data reconstructed from a PCA and the percentage of variance reproduced thank's to the PCA.

Data_preparation.write_dat — Method

write_dat(matrix,PATHTOSAVE,NAME ; more=[""], overwrite=false)

Write matrix in a .dat file. Each column of the matrix is writen as new column in the file. Delimiters are spaces by default. The option overwrite is false by default. If want to overwrite file with the same name as given in input, change it to true. If want to add more entries on the header of the dat, use the option 'more'. Will be added as commentaries with a # at the begining of the line.

Data_preparation.write_fits — Method

write_fits(fitstocopy::String,newname::String,pathtosave::String,datatosave,datadim,blank ; finished=true)

Create a fits with its own header based on 'fitstocopy', and with the name 'newfits'. DO NOT ADD THE EXTENSION (e.g ".fits") FOR THE NAME OF THE NEWFITS). Need to give the blanking value. The fits can be constructed by data with two dimensions or three dimensions. Variable 'finished' is by default true, so the function will by default close the new fits file at the end of the execution of this function. If 'finished' is false, the function will not close the fits file, allowing it to be modified again in another function (createcvifits for example). If want to add more entries on the header, use the option 'more'. Alternate between the name printed in the header and the value. Example, if want to add an entry for 'imsize' equals to 50, and an entry for 'powerlaw' equals to -3, write : more=["imsize","50","powerlaw","-3"]. The option overwrite is false by default. If want to overwrite file with the same name as given in input, change it to true.

Data_preparation.write_pca_fits — Method

write_pca_fits(pc,varpercent,fitstocopy::String,newname::String,data,datadim,blank)

Create a fits used for pca reconstructed data. Contains on its header the number of pc used and the percentage of variance explained. Need a fits to copy its header entries.

PCA

Functionforpca.convergence_moments_byintegration — Method

    convergence_moments_byintegration(data2D,integration_range,pcnumber,velocity_increment,M,Yt,datadim,velocity_vector,IM)

Return the I-th moment order of the integrated by sections spectras differences, from arrays reconstructed with N PC and arrays without PCA. Also return the new velocity vector where each value is centered on the integrated sections. The function integrate by section (the integration_range) spectra from reconstructed and non-reconstructed data, then compute the given moment order of their differences. Data should be in 2 dimensions (PV cube), and datadim should refers to the dimension of the 3D data cube (PPV). Pcnumber should be a vector with all the number of PC you want (example Vector(1:10)).

Functionforpca.cv_convergence — Method

    cv_convergence(pcmax,M,Yt,velocity_vector,datadim)

Return a cube reconstructed with different number of PC, its first and its second velocity moment order. The cube considered is the one processed by PCA, that gave M and Yt as a result. The arrays in output are in 3D, where the third dimensions gives results for different number of PC used. Datadim should be the dimension of the 3D initial PPV cube. M is a PCA type, produced by the function pca. Yt is a vector or a matrix where each column gives the principal components of an observation.

Functionforpca.intensity_moments_specific_channels — Method

    intensity_moments_specific_channels(arr_path,channels,datadim)

Will compute the first, second, third and fourth, moments order of the mean of given channels in a data cube. The cube is not given as input but only the path where it is saved as a 'bin' temporary file (saving RAM). Datadim refer to the dimension of the 3D data cube (the original data cube).

Functionforpca.intensity_moments_specific_channels_withPCA — Method

    intensity_moments_specific_channels_withPCA(M,Yt_path::String,pcnumber,canals,datadim)

Will compute the first, second, third and fourth moments order of the mean of specific canals in the data reconstructed with different number of PC to check if they are converging. Datadim refer to the dimension of the 3D data cube (the original data cube). M is a PCA type, produced by the function pca. Yt_path is the path where Yt is saved as a 'bin' temporary fil. Yt is a vector or a matrix where each column gives the principal components of an observation. M and Yt are obtained with the function pca. Moment order can be any Int from 1 to 4. Return one 1D array.

Functionforpca.intensity_moments_specific_channels_withPCASVDRAND — Method

    intensity_moments_specific_channels_withPCASVDRAND(svdobject,pcnumber,channels,p)

Same as intensitymomentsspecificchannelswithPCA but using a Principal Component Analysis based on random SVD.

Functionforpca.multiple_moment! — Method

    multiple_moment!

Same as multiple_moment but on pre-allocated data.

Functionforpca.multiple_moment — Method

    multiple_moment(velvector,data_dimension,M,Yt,pcmax)

Compute first and second velocity moment order of multiple data cube. The data cubes are reconstructed from PCA with different number of PC. M is a PCA type, produced by the function pca. Yt is a vector or a matrix where each column gives the principal components of an observation.

Functionforpca.multiplepca — Method

    multiplepca(M,Yt,pcnumber,datadim)

Do multiple PCA with different number of PC on the same data. Will return a 3D array, with pixel position in the first dimension, velocity in the second dimension, and the data reconstructed with a specific number of pc in the third dimension. Pcnumber should be a vector with all the number of PC you want (example Vector(1:10)). M is a PCA type, produced by the function pca. Yt is a vector or a matrix where each column gives the principal components of an observation.

Functionforpca.pca — Method

    pca(data2D,pc::Integer,path::String ; percent=1.0)

Produce a Principal Component Analysis on a data, using 'pc' number of principal component. 'percent' gives the wanted percentage of variance reconstructed (default=1). 'path' is the path where results will be saved (in 'bin' temporary files, named as mmapYt.bin, mmapMproj.bin and mmapMmean.bin) Data are converted in an Array{Float64,2} to do the PCA ( type as Union{Missing} will produce an error). Data should be given in 2D (pixels per velocities). Will return the path to the projections of the data2D on the PC axis, the mean of these projections, the path to the matrix composed of the principal components in columns and the percentage of explained variance. These results will allow to reconstruct the cube with a given number of PC (using pcanomorecalc).

Functionforpca.pca — Method

    pca(data2D,pc::Integer ; percent=1.0)

Produce a Principal Component Analysis on a data, using 'pc' number of principal component. 'percent' gives the wanted percentage of variance reconstructed (default=1). Data are converted in an Array{Float64,2} to do the PCA ( type as Union{Missing} will produce an error). Data should be given in 2D (pixels per velocities). Will return the pca fit M (a type created by the MultivariateStats package), the matrix composed of the principal components in columns, the percentage of explained variance and the data reconstructed with 'pc' number of principal component.

Functionforpca.pca_nomorecalc — Method

    pca_nomorecalc(M,Yt,pcfirst,pclast,Datadim)

Reconstruct data with a certain number of PC, without redo all of the PCA calculation. Necessary to do a first pca calculation to run this one, because needed M (a new type related to PCA constructed by MultivariateStats package) and Yt (matrix with each column a PC). Datadim should be a tuple with two values (PV cube, 2D) or three (PPV cube, 3D).

Functionforpca.pca_nomorecalc — Method

    pca_nomorecalc(Mproj_path::String,Yt_path::String,Mmean_path::String,pcfirst,pclast,datadim)

Same as pca_nomorecalc(M,Yt,pcfirst,pclast,Datadim) but using here the path to the temporary files containing the informations (M and Yt). To be used after pca(data2D,pc::Integer,path::String ; percent=1.0) in order to produce a reconstructed cube.

Functionforpca.variance_evolution — Method

    variance_evolution(M,threshold)

Return an Array with the numbers of PC explaining a percentage of the data variance tinier (or equal) than the threshold given in input.

SWO

Spectralwindowopti.convoptiwind — Method

convoptiwind(cubesource,DATADIMENSION,VELOCITYVECTOR,NOISECAN,BLANK,RANGE,PATHTOSAVE)

DATADIMENSION should be 2D : PV

Spectralwindowopti.optiwind — Method

optiwind(cube,DATADIMENSION,VELOCITYVECTOR,NOISECAN,BLANK,RANGE)

INPUT DESCRIPTION Method of window optimisation. Search for the limits of the signals in each spectrum of the cube. The cube given as input should be in 2Dimensions, with spectrum ordonnated by row. A first CV computation on each spectra should also be given as input : cvmap. Cvmap is in 1D, with each position corresponding to the spectrum of the same position in the 2D cube. The velocity vector should also be given (obtained by Datapreparation.readfits_ppv). NOISECAN corresponds to a Vector{Int64} of 2 entries, where each entry corresponds to a position in the Velocity Vector of a noise cannal : they have to englobe only noise. Used to compute the noise rms ; ESSENTIAL IN THE COMPUTATION. The last value RANGE is the size of the increment used during the computation : lower it is and longer the code will be.

METHOD A first initialisation of a mask of same size than the input cube, with 0 everywhere. For each spectra, compute the intensity area by velocity increments from each side of its CV value. When the integration is lower than sqrt{N}*NoiseRMS, the limit is kept (with N the number of velocity cannals used for the integration and NoiseRMS the noise rms). When the two limits are obtained, change values of the mask from 0 to 1 inside these limits for each spectra. Last, multiplication of the mask to the cube.

OUTPUT DESCRIPTION Return the cube masked and the mean of the noise accross the map (one sigma).

CVI

Functionforcvi.construct_cvimap! — Method

construct_cvimap!(xyarr,Lag::Int64,nangle,mapdim,cvi_averaged::Array{Union{Missing, Float64}},cvi_allangle::Array{Union{Missing,Float64},3}; diff="relative",keepmissing=true)

Will construct a cvi map using preallocated array. Need an array preallocated for the storage of the average on all cvi angle (cviaveraged), and another for the cvi calculated with all angle and lag (cviallangle). The cvi map is constructed by taking the mean of all rotations of the cv increment calculation at each pixel. The Lag is the increment. xyarr have to be in 2D (pixel*pixel). Mapdim is the dimension of your 2Dmap. The differences can be absolute or relative. Nangle refers to the number of rotations of the cvmap: a higher value will compute a higher number of differences pixel per pixel, so a higher number of directions.

Functionforcvi.construct_cvimap! — Method

construct_cvimap!(xyarr,Lag::Vector{Int64},nangle,mapdim,cvi_averaged_alllag::Array{Union{Missing, Float64},2},cvi_allangle_alllag::Array{Union{Missing, Float64}, 3},cvi_allangle::Array{Union{Missing,Float64},3}; diff="relative",keepmissing=true)

Will construct a cvi map using preallocated array. Need an array preallocated for the storage of the average on all cvi angle (cvimean), and another for the cvi calculated with all angle and lag (cviallanglealllag). The cvi map is constructed by taking the mean of all rotations of the cv increment calculation at each pixel. The Lag is the increment. xyarr have to be in 2D (pixel*pixel). Mapdim is the dimension of your 2Dmap. The differences can be absolute or relative. Nangle refers to the number of rotations of the cvmap: a higher value will compute a higher number of differences pixel per pixel, so a higher number of directions.

Functionforcvi.construct_cvimap — Method

construct_cvimap(cvmap,Lag::Int64,nangle,mapdim; diff="relative",keepmissing=true)

Same as construct_cvimap if Lag is a Int64 of one lag. Mapdim is the dimension of your 2Dmap. The differences can be absolute or relative. Nangle refers to the number of rotations of the cvmap: a higher value will compute a higher number of differences pixel per pixel, so a higher number of directions.

Functionforcvi.construct_cvimap — Method

construct_cvimap(cvmap,Lag::Vector{Int64},nangle,mapdim; diff="relative",keepmissing=true,BLANK=-1000)

Construct a Centroid Velocity Increment map based on a 'cvmap'. The cvi map is constructed by taking the mean of all rotations of the Centroid Velocity Increment calculation at each pixel. The Lag is the increment. xyarr have to be in 2D (pixel*pixel). Mapdim is the dimension of your 2Dmap. The differences can be absolute or relative. Nangle refers to the number of rotations of the cvmap: a higher value will compute a higher number of differences pixel per pixel, so a higher number of directions.

Functionforcvi.construct_cvmap — Method

construct_cvmap(xyarr,velvector)

Will construct a cv map by calculating the first velocity moment order of the array and replace the inf values by NaN. xyarr have to be in 2D (pixel*pixel), while velvector have to be in 1D (velocities).

Functionforcvi.cv_increment! — Method

cv_increment!(xyarr,Lag::Int64,blank,cvimap_lag::Array{Union{Missing,Float64},2})

Same as (xyarr,Lag::Int64,nangle; diff="relative",periodic=false) but on preallocated array (array produced outside this function). To produce the array, use : cvi_allangle = Array{Union{Missing,Float64},3}(undef,DataDimension[1],DataDimension[2],nangle).

Functionforcvi.cv_increment! — Method

cv_increment!(xyarr,Lag::Vector{Int64},nangle,cvi_allangle_alllag::Array{Union{Missing, Float64}, 3},cvi_allangle ; diff="relative",periodic=false)

Same as cvincrement(xyarr,Lag::Vector{Int64},nangle; diff="relative",periodic=false) but on preallocated array (array produced outside this function). To produce the array, use : cviallangle = Array{Union{Missing,Float64},3}(undef,DataDimension[1],DataDimension[2],nangle) and cviallanglealllag = Array{Union{Missing,Float64},3}(undef,DataDimension[1]*DataDimension[2],nangle,size(Lag)[1])

Functionforcvi.cv_increment — Method

cv_increment(xyarr,Lag::Int64,nangle; diff="relative",periodic=false)

Compute the centroid velocity increment of xyarr at one Lag. Nangle is the number of angle using to compute the differences (it's a value in the parameter file, equal to 192). Diff (default relative) is for differences between two pixels : absolute or relative. Periodic=true (default=false) is for working on periodic data (from simulations like fbm). The returned array will have the first two dimensions equal to the size of the map, and the third dimension is the cvi computed at each angle.

Functionforcvi.cv_increment — Method

cv_increment(xyarr,Lag::Vector{Int64},nangle; diff="relative",periodic=false)

Compute the centroid velocity increment of xyarr at multiple Lag values. Nangle is the number of angle using to compute the differences (it's a value in the parameter file, equal to 192). Diff (default relative) is for differences between two pixels : absolute or relative. Periodic=true (default=false) is for working on periodic data (from simulations like fbm). The returned array will have the first dimension equal to the size of the map (pixel square), the second dimension is the cvi computed at each angle, and the third dimension is for each value of Lag.

Functionforcvi.delete_aberations — Method

delete_aberations(arr1,arr2)

Replace values in arr1 which are higher or lower than values in arr2 by missing. Don't use an arr1 with missing values !

Functionforcvi.moment_four — Method

moment_four(yarr,xarr)

Return the fourth moment order of xarr weighted by yarr.

Functionforcvi.moment_one — Method

moment_one(yarr,xarr)

Compute the first moment order of the array xarr weighted by yarr. Centroid = int(T*v)dv/int(T)dv

Functionforcvi.moment_one_field — Method

moment_one_field(arr,velvector)

Return the first moment order of all pixels in an entire field, weighted by velvector. Arr can be a 3D array (ppv) or a 2D array (pv)

Functionforcvi.moment_three — Method

moment_three(yarr,xarr)

Return the third moment order of xarr weighted by yarr.

Functionforcvi.moment_two — Method

moment_two(yarr,xarr)

Return the second moment order of xarr weighted by yarr.

Functionforcvi.moment_two_field — Method

moment_two_field(arr,velvector)

Return the order 2 moment of an entire field.

Functionforcvi.multiple_moment — Method

multiple_moment(cube::Array{Float64,3},velvector)

Return the first or second velocity moment order of a 3D array. The 3D array in input have its first dimension giving the pixel position, the second dimension giving spectra, and the third dimension giving other data with the same dimensions. This function was built in order to obtain in the third dimensions multiple centroid velocity maps with data reconstructed from different number of principal component (see Principal Component Analysis)

Graphical representations

Graphic.StcFct — Method

StcFct()

Plot structure functions of orders p in function of the structure function of order 3. If add is true, the plot will be added on the actual plot. The figure can be saved with the number of PC used during the process in the name by changing pcfinal.

Graphic.StcFctExponent — Method

StcFctExponent(zeta,third_order_zeta,OrderP,xlim,ylim,labs,data_name_title,directories_name; add=false,save=true,pcfinal=0)

Plot the exponent of the structure function power-law fit in function of the order p of the structure function. If add is true, the plot will be added on the actual plot. Also add a y=1/3x line on the plot. The figure can be saved with the number of PC used during the process in the name by changing pcfinal.

Graphic.StcFctExponentWithError — Method

StcFctExponentWithError()

Plot the exponent of the structure function power-law fit in function of the order p of the structure function. Compute and plot also the error of the fit. If add is true, the plot will be added on the actual plot. Also add a y=1/3x line on the plot. The figure can be saved with the number of PC used during the process in the name by changing pcfinal.

Graphic.StcFctWithFit — Method

StcFctWithFit()

Plot structure functions of orders p in function of the structure function of order 3 with the fit used to obtain the power-law exponent. If add is true, the plot will be added on the actual plot. The figure can be saved with the number of PC used during the process in the name by changing pcfinal.

Graphic.animate_ppvcube — Method

animate_ppvcube(cube,data_name)

Plot a heatmap of a 3D cube along third dimension, and save it as a gif file.

Graphic.animate_spectra — Method

animate_spectra(arr,positionnumber,direction)

Make an animation (.gif) plotting all the spectrum of a line of pixels (dims=1) or a row of pixels (dims=2). Array has to be 3D. Used the array, the position of the line or the row, and the direction (line=1 or row=2).

Graphic.checkwindowopti — Method

Graphic.contourmap — Method

contourmap(arr,arraytocontour,lev,xvec,yvec)

Plot the data arr in heatmap type and contours in front at levels=lev. Contours can be provided from another data. Arr has to be 2D.

Graphic.corner_cvimap — Method

corner_cvimap(array1,array2,datadim,labelarray1,labelarray2)

Make a cornerplot between each element of the third dimension of array2, and also compare with array1. Array1 has to be 1D (cvi pixel values), and array2 has to be 2D (cvi pixel values and different cvi maps obtained (with different PC for example)). Labelarray1 has to be 1 String, Labelarray2 has to be an array os String, with as elements as the third dimension of array2. ["dollars(ix) PC" for ix=2:4] to label the named of the third dimension of array2 has number of PC used

Graphic.cvi_pdf — Method

cvi_pdf(data,para0,bin,ylim1,ylim2,alph,Lag,data_name_title,legend ; add=false)

Plot the pdf of the cvi, WIP : and add a gaussian fit on the data. By default, will do a new plot. If add=true, will add this plot on the last figure created.

Graphic.cvi_pdf_norm — Method

cvi_pdf_norm(data,sigm,bin,ylim1,ylim2,alph,Lag,data_name_title,legend ; add=false)

Plot the pdf of the cvi, WIP : and add a gaussian fit on the data. By default, will do a new plot. If add=true, will add this plot on the last figure created.

Graphic.heatmap_compare_multiplepc — Method

compare_rawpc(xvec1,yvec1,array1,dimens,xvec2,yvec2,arraypc,title1::String,saveall::Bool,field)

Plot multiple files, each with two heatmap : the first is a heatmap of the array1 (same for all file), the second a heatmap of the arraypc which change for each file in order to plot all of its third dimension (Data reconstructed with different number of PC for example). Array1 has to be 2D (PixelPixel), arraypc has to be 3D (PixelPixel*PC). The subplots of each files have the same colorscale.

Graphic.heatmap_interacted — Method

heatmap_interacted(data,delta_xvec,delta_yvec)

Will produce a heatmap plot of a 2d dataset, with colorscale changing by an interaction of the user on a slider.

Graphic.heatmap_subplot_same — Method

heatmap_subplot_same(arr,j,velocity_range)

Made a subplot in a heatmap form, using the one data but plotting different velocity canals (velocity_range). J is the position of the first plot.

Graphic.heatmap_subplot_two — Method

heatmap_subplot_two(xvec1,yvec1,array1,xvec2,yvec2,array2,lim1,lim2,title1,title2,save,field)

Made a subplot in a heatmap form, using two different data (array1 and array2) with a defined colorbar scale (lim1 and lim2), their x and y coordinates (xvec1,yvec1 and xvec2,yvec2) a title for each plot (title1 and title 2), if the plot need to be saved (save) and where.

Graphic.ifdiff — Method

ifdiff(spectratokeep,array1,reconstructed_array,velocity_vector,datadim,titl,nbrow,nbcol)

Plot the spectra associate with absolute differences in moment calculation (N moment order or CV) higher than a threshold between two data cube. Plot also the differences of the spectra. Arraymoment and reconstructedmoment are two 2D arrays, while array1 and reconstructed_array are 3D data cube (PPV).

Graphic.ifdiff_cvmap — Method

ifdiff_cvmap(cvmap,specdiff_map,xvec,yvec,lim1,lim2,titl)

Add the specdiffmap on the plot of the cvmap . Specdiffmap can be a map where each pixel missing is a spectrq giving a difference in the CV calculation > to a threshold, qnd the others pixels are equal to 0.

Graphic.meanpc — Method

meanpc

Plot two images in one plot : the first is the mean of all the spectrums of a data (using directly the rawdat, will be meaned in the function), the second is the data reconstructed with N PC (npc) obtained by a pca method on these data, using M which is a PCA object type introduced in MultivariateStats package. Need the scales of the color bars (scale1 and scale2), the titles of each plot (title1 and title2), a boolean to know if the plot need to be saved (save), the name of the save if needed (filename), and if the plots needs to be display. Arr can be 2D array (PV) or a 3D array (PPV).

Graphic.mom_conv — Method

mom_conv(momarray,array_specavr,velocity_vector,pcnumb,momentorder,nbrow,nbcol,velocity_range)

Same as the function moment_byintegration, but all plotted in different subplot in the same file. Pcnumb is a vector with all PC number you work with. Momentorder should be equal to the moment number you want to plot (used for the title).

Graphic.moment_byintegration — Method

moment_byintegration(momarray,pcnumb,velocity_vector,momentorder)

Plot the 3rd or 4th moment order of the differences of integrated by sections spectras between original data (without PCA) and reconstructed data with PCA. Pcnumb is a vector with all PC number you work with. Momentorder should be equal to the moment number you want to plot (used for the title).

Graphic.moment_multiplepc — Method

moment_multiplepc(moment_multiplepc,pcnumb,nbrow,nbcol,titl,limx,limy;disp=false)

Display scatter subplots comparing moment maps (N moment order or CV) obtain with different number of PC. NPC is compared with (N-1)PC. A linear function y=x is also plotted. Moment_multiplepc is a 3D array where the third dimension gives different maps calculated with a number of PC different.

Graphic.moment_multiplepc_withraw — Method

moment_multiplepc_withraw(moment_multiplepc,pcnumb,array1,nbrow,nbcol,titl)

Same as moment_multiplepc but compare each moment maps obtain with different number of PC with the original data (without PC). A linear function y=x is also plotted.

Graphic.moment_specific_canals — Method

moment_specific_canals(moment_pca,moment_nopca,data_name,pcmax)

Plot a file with 4 subplots, each showing a moment calculation on the mean intensity of specific canals (generally associated with noise). 'momentpca' are the moments calculated on data reconstructed from PCA, and 'momentnopca' the moments calculated on the original data set.

Graphic.pcacompwo — Method

Graphic.pixels_averaged_spectrum — Method

pixels_averaged_spectrum(arr,arr2,nbrow,nbcol,titl,pixel_range::Tuple,blank,vel1)

Plot the averaged spectrums of two cube arr and arr2. The average is doing on a square of (pixelrange) by (pixelrange) pixels.

Graphic.pixels_averaged_spectrum — Method

pixels_averaged_spectrum(arr,nbrow,nbcol,titl,pixel_range:Tuple,blank,vel1)

Plot the averaged spectrums of a data. The average is doing on a square of (pixelrange[1]) by (pixelrange[2]) pixels.

Graphic.pixels_averaged_spectrum_twodata — Method

pixels_averaged_spectrum_twodata(arr,reconstructed_array,nbrow,nbcol,titl,pixel_range,blank,vel1,vel2)

Plot the averaged spectrums of two fits : the rawdata and the data reconstructed from a PCA applied to these rawdata, with two velocity vectors different. The average is doing on a square of (pixelrange) by (pixelrange) pixels.

Graphic.ploptiwind — Method

Graphic.plpc — Method

    plpc(Yt,pc)

Plot some principal components (PC) of data analysed by a PCA. Yt is the transformation of the raw data into principal components.

Graphic.pratio — Method

pratio(M::PCA,ylog::Bool,pc,titl)

Plot the explained percentage of the data per principal component. If ylog is True, the yaxis will be plotted in log. Pc arguments are used to name the plot.

Graphic.region_spectrum — Method

region_spectrum

Plot spectra from a specific region

Graphic.region_spectrum_twodata — Method

region_spectrum_twodata

Graphic.sci_not — Method

sci_not(x,ndec)

Used to transform axis x labelling in scientific notation with ndec the number of decimals.

Function for fBms data based

Functionforfbm.fbm2D — Method

fbm2D(imsize,powerlaw ; disp = false)

Construct a 2D fbm map of size imsize*imsize, following a power spectrum law equal to powerlaw. Can plot the map (disp = true).

Functionforfbm.power_spectra — Method

power_spectra(arr,karr,imsize ; fitted=true)

Compute the power spectrum of arr, print and plot it. karr is the frequencies array (pixel^-1). Also add a powerlaw fit by default.

Functionforfbm.ppv_from_fbm — Method

ppv_from_fbm(imsize,powerlaw,xarray,sig,stdev,tmax)

Compute a ppv cube, with two gaussian components centered on the values obtained in 2D fbm maps minus the mean of the xarray (== velocity array). A noise with a standard deviation of stdev is added.

Functionforfbm.ppv_from_fbm2 — Method

ppv_from_fbm2(imsize,powerlaw,xarray,sig,stdev,tmax)

Compute a ppv cube, with one gaussian components centered on the values obtained in 2D fbm maps minus the mean of the xarray (== velocity array). A noise with a standard deviation of stdev is added.

Functionforfbm.ppv_from_fbm4 — Method

ppv_from_fbm4(imsize,powerlaw,xarray,sig,stdev,tmax)

Compute a ppv cube, with 4 gaussian components of maxima equal to the values obtained in 2D fbm maps minus the mean of the xarray (== velocity array). A noise with a standard deviation of stdev is added. Xarray is the velocity axis wanted (write it as : -20:01:20 for example) xinc is an array (3 of size) given the increments for each components. sig is an array (3 of size) given the dispersions. A random value will also be added.

Functions for Data Analysis

Data_analysis.calcmetric — Method

calcmetric(mom1,mom2,mom3,mom4,dv)

Calculate the following metric on each values of mom1,2,3 and 4: sqrt((mom1/dv)2+(mom2/dv)2+(mom3)2+(mom4-3)2)

Data_analysis.find_indices — Method

find_indices(specdiff,arra ; dims = size(arra))

Return an array of the same dimension than arra where the indices in specdiff are equal to 1000, and others indices are equal to 0. Dims indicates the dimensions of the returned array (by default same as in input). Use this if some dimensions should not be used.

Data_analysis.find_threshold — Method

find_threshold(array1,array2,thresh)

Return an array with each value being the index where abs(array1[indice]-array2[indice]) is greater than a threshold.

Data_analysis.fourmoments — Method

fourmoments(cube;dim=2)

Compute the first four moments order of a given cube of dimension 'dim'. If 2D PV cube, will compute the moments on each row. Return the four moments. They are Vector of dimension=size(cube)[1]

Data_analysis.generate_gaussian — Method

generate_gaussian(xarray,mu,sig)

Produce a gaussian and a normalized gaussian, following mu and sigma.

Data_analysis.integration_bysection — Method

integration_bysection(data2D,integration_range,datadim,increment)

Compute an integration of spectra with a given increment, for each spectra in a map. Return an array with the first dimension giving a different spectra, and the second dimension giving the integration sections.

Data_analysis.rms_analytic — Method

rms_analytic(yarr,rms,xarr)

Return the uncertainty of the first moment order in velocity, and some part of the analytical expression. Yarr represents the emission part of the data, xarr the velocity vector.

Data_analysis.rms_analytic_field — Method

rms_analytic_field(array1,xarr,noise_canals)

Return the uncertainty of the first moment order in velocity of a field, and some part of the analytical expression. Xarr represents the velocity vector. Array1 is a 2D data cube, with each row a new pixel and the columns forms the spectra.

Data_analysis.snr — Method

snr(yarr,noise_canals)

Return the SNR of a signal (spectra for example). yarr have to be a vector.

Data_analysis.snr_allfield — Method

snr_allfield(arr,noise_canals)

Return the mean of the SNR calculated on several spectra. Arr can be 2D (PV) or 3D (PPV) but will be faster if 2D. Better to avoid missing values in your data.

WORK IN PROGRESS : Function for Structure functions computations

Structure_fct.construct_fctstruct! — Method

construct_fctstruct!

Construct a matrix with the values of the structure functions at every given lag and orders (lag in columns and orders in lines). Need a preallocated array.

Structure_fct.construct_fctstruct — Method

construct_fctstruct

Construct a matrix with the values of the structure functions at every given lag and orders (lag in lines and orders in columns).

Structure_fct.fct_struct — Method

fct_struct(data,datapdf,order,bin)

Compute the structure function of a data at a specific order.

Structure_fct.fit_fctstruct — Method

fit_fctstruct(xdata,ydata,y0,confidinterv ; confinterv=true)

Fit the model ydata=A*(xdata)^B. Can return the confidence interval too of the model. Return in first B, then A.

Structure_fct.pdf_normed — Method

pdf_normed(data,bin,bound::Tuple)

Compute the PDF of the data given, and normalize it by centered at 0 mean, a standard deviation of unity and an area=1.

Structure_fct.xhi_fct_p — Method

xhi_fct_p(pvec,struct_lag,lagvec,y0,confidinterv)

Fit the model y=A*(x)^B for multiple order and lag of structure functions.